Factor out address import code to separate module.
